pyproject.toml
[project]
name = "omninomicon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sentence-transformers>=5.2.0",
    "sqlmodel>=0.0.31",
]
